Supported CDN urls for media & files
This is an initial spike into handling CDN urls for media and files.

In order to use this you should update the config to look like:

```
{
  "urls": {
    "media": "https://storage.ghost.is/c/12/34/1234567890",
    "files": "https://storage.ghost.is/c/12/34/1234567890"
  }
}
```

And now all media and file urls that are read from the DB will be
pointing to the CDN, and all media and file urls that are pointing to
the CDN will be stored as `__GHOST_URL__` in the DB.

Things to note:

 - This doesn't work for mobiledoc posts yet, but that shouldn't affect
   testing your own sites on staging, they will all use lexical.

 - You can get some weird unsaved changes modal behaviour when the urls
   config has been set but the storage adapter has not been updated to
   return the same urls, so make sure to update storage adapters.

This depends on the 5.0.0-rc.1 release of @tryghost/url-utils. The
source code for this is available on the PRO-1524 branch here:
 - https://github.com/TryGhost/SDK/tree/PRO-1524/packages/url-utils